viot/pyproject.toml
[tool.poetry]
name = "viot"
version = "0.1.0"
description = "Viot"
authors = ["tuanvumaihuynh <tuanvux01@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app", from = "."}
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.30.6"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.4.0"
uvloop = "^0.19.0"
msgspec = "^0.18.6"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
redis = {extras = ["hiredis"], version = "^5.0.7"}
python-slugify = "^8.0.4"
bcrypt = "^4.2.0"
yagmail = "^0.15.293"
jinja2 = "^3.1.4"
alembic = "^1.13.2"
celery = "^5.4.0"
pyjwt = "^2.9.0"
httptools = "^0.6.1"
injector = "^0.22.0"
classy-fastapi = "^0.6.1"
flower = "^2.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.7"
pre-commit = "^3.8.0"
pytest-asyncio = "^0.23.8"
pytest-xdist = "^3.6.1"
coverage = "^7.6.1"
httpx = "^0.27.0"
mypy = "^1.11.2"
types-redis = "^4.6.0.20240819"
celery-types = "^0.22.0"
faker = "^28.4.1"
testcontainers = "^4.8.1"


[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = [
    "app/database",
    "app/common",
    "app/module/**",
]


[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 80

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e
]

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]